+++ apps/web/app/_features/top/index.ts
@@ -1,1 +1,1 @@
-export * from "./WorkExperience";
+export * from "./WorkExperienceList";
